Task/requirements.txt
python=3.11.5
django=5.0.3
djangorestframework

#adding user from django admin panel or clients page